user-service/requirements.txt
grpcio==1.51.3
grpcio-tools==1.51.3
protobuf==4.22.1
psycopg2-binary==2.9.6